one (simple) dev script to rule them all
its just simpler
