+++ CDS/scripts/js_utils/package.json
@@ -22,6 +22,7 @@
     "mocha-test": "STAGE=test mocha test/**/*.js test/*.js"
   },
   "devDependencies": {
+    "argparse": "^1.0.9",
     "chai": "^3.5.0",
     "eslint": "^3.12.2",
     "jshint": "^2.9.4",
